+++ src/contexts/index.ts
@@ -1,6 +1,6 @@
 export { EdgesContext, useEdgesProperties } from './EdgesContext';
 export { GraphContext, useGraphProperties } from './GraphContext';
 export { LayoutContext, useLayoutProperties } from './LayoutContext';
-export { ModalsContext, useModals } from './ModalsContext';
 export { NodesContext, useNodeProperties } from './NodesContext';
+export { PopupsContext, useModals, useToasts } from './PopupsContext';
 export { SettingsContext, useSettings } from './SettingsContext';